move generated code to its own folder
