+++ lib/thrift/geocoder.rb
