+++ lib/thrift/geocoder_types.rb
