+++ lib/twofishes.rb
@@ -1,8 +1,8 @@
 require 'twofishes/version'
 require 'twofishes/configuration'
 require 'thrift_client'
-require 'twofishes/geocoder_types'
-require 'twofishes/geocoder'
+require 'thrift/geocoder_types'
+require 'thrift/geocoder'
 require 'twofishes/errors'
 require 'twofishes/result'
 require 'twofishes/underscored'
